Add files via upload
This is the single stack program created using OpenXTalk.  The program is billing software for solo public defenders in Oregon but should work for public defenders in any state in the United States as long as the solo public defense attorney only gets funding from one source (i.e.: the state agency in charge of public defense).  The code is distributed here for free under MIT license so modify it to suit your needs.  This is the only free legal billing software I know of.  The compiled (application program) is available on my website https://www.pdxdefense.com/ or https://www.aboutwes.com